feat: refactor module structure and add classification metrics
