+++ ifri_mini_ml_lib/association_rules/__init__.py
@@ -1,11 +1,9 @@
 from .apriori import Apriori
 from .eclat import Eclat
 from .fp_growth import FPGrowth
-from .utils import DataAdapter
 
 __all__ = [
     "Apriori",
     "Eclat",
     "FPGrowth",
-    "DataAdapter"
 ]
