+++ ifri_mini_ml_lib/preprocessing/dimensionality_reduction/__init__.py
@@ -1,5 +1,5 @@
 from .pca import PCA
-from .t_sne import TSNE
+from .tsne import TSNE
 
 __all__ = [
     "PCA",
