+++ ifri_mini_ml_lib/preprocessing/preparation/__init__.py
@@ -1,8 +1,8 @@
-from .min_max_scaler import MinMaxScaler
+from .encoding import CategoricalEncoder
 from .missing_value_handler import MissingValueHandler
-from .standard_scaler import StandardScaler
-from .categorical_encoder import CategoricalEncoder
-from .data_splitter import DataSplitter
+from .splitting import DataSplitter
+from .scaler.min_max_scaler import MinMaxScaler
+from .scaler.standard_scaler import StandardScaler
 
 __all__ = [
     "MinMaxScaler",
