+++ ifri_mini_ml_lib/preprocessing/preparation/scaler/min_max_scaler.py
