+++ ifri_mini_ml_lib/preprocessing/preparation/scaler/standard_scaler.py
