+++ ifri_mini_ml_lib/regression/polynomial_regression.py
@@ -2,7 +2,7 @@ from typing import Union, List
 
 import numpy as np
 
-from .linear import LinearRegression
+from .linear_regression import LinearRegression
 
 
 class PolynomialRegression:
